borg: Replace package.el with borg
Borg is a bare-bones package manager for Emacs packages. It provides
only a few essential features and should be combined with other tools
such as Magit, epkg, use-package, and auto-compile.

Borg assimilates packages into the ~/.emacs.d repository as Git
submodules. An assimilated package is called a drone and a borg-based
~/.emacs.d repository is called a collective.