eglot: Assimilate with dependencies
eglot is a more lightweight client for Language Server Protocol
servers as lsp-mode.

Its design is intended to be included in emacs, and so its
dependencies are emacs core-packages, which are not yet in the current
release (flymake 1.0.2 and jsonrpc 1.0.7).
Since these packages are bundled inside the emacs-repository, their
installation through borg uses copies hosted elsewhere.