evil: Differentiate between tab and C-i
Tab and ctrl-i are generally the same thing, especially in the
terminal. While Emacs tab to indent the current line correctly is
really handy, as is the cycling through headings in org-mode, I also
miss the jump-forward on ctrl-i that is the counterpart to vim's
ctrl-o.

So instead differentiate between tab and C-i, so that they can do
different things.